Restructure components dir to have sub-dirs
