+++ src/components/navigation/NavPage.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import PropTypes from "prop-types";
-import { AppContext } from "../contexts/AppContext";
+import { AppContext } from "../../contexts/AppContext";
 
 function NavPage(props) {
   const { id, children } = props;
