+++ src/components/navigation/Navigation.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { AppContext } from "../contexts/AppContext";
+import { AppContext } from "../../contexts/AppContext";
 
 function Navigation() {
   const { navPages, activeNavPage, setActiveNavPage, navSearchButtonRef } =
