+++ src/components/search/SearchResultItem.jsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState, useEffect } from "react";
-import { iiifUrl } from "../utils";
-import { AppContext } from "../contexts/AppContext";
+import { iiifUrl } from "../../utils";
+import { AppContext } from "../../contexts/AppContext";
 import PropTypes from "prop-types";
 
 /**
